FieldmarkImportDialog component
