+++ src/main/webapp/ui/src/Inventory/components/CreateNew.js
@@ -4,7 +4,7 @@ import React, { type Node, useContext, type ComponentType } from "react";
 import useStores from "../../stores/use-stores";
 import NavigateContext from "../../stores/contexts/Navigate";
 import { UserCancelledAction } from "../../util/error";
-import { styled, ThemeProvider } from "@mui/material/styles";
+import { styled } from "@mui/material/styles";
 import Menu from "@mui/material/Menu";
 import NewMenuItem from "../../eln/gallery/components/NewMenuItem";
 import RecordTypeIcon from "../../components/RecordTypeIcon";
@@ -12,46 +12,11 @@ import Divider from "@mui/material/Divider";
 import Button from "@mui/material/Button";
 import Box from "@mui/material/Box";
 import { observer } from "mobx-react-lite";
-import Dialog from "@mui/material/Dialog";
-import createAccentedTheme from "../../accentedTheme";
-import Toolbar from "@mui/material/Toolbar";
-import AppBar from "@mui/material/AppBar";
-import AccessibilityTips from "../../components/AccessibilityTips";
-import HelpLinkIcon from "../../components/HelpLinkIcon";
-import DialogContent from "@mui/material/DialogContent";
-import Grid from "@mui/material/Grid";
-import Link from "@mui/material/Link";
-import Typography from "@mui/material/Typography";
 import FieldmarkIcon from "../../eln/apps/icons/fieldmark.svg";
 import CardMedia from "@mui/material/CardMedia";
-
-export const FIELDMARK_COLOR = {
-  main: {
-    hue: 82,
-    saturation: 80,
-    lightness: 33,
-  },
-  darker: {
-    hue: 82,
-    saturation: 80,
-    lightness: 22,
-  },
-  contrastText: {
-    hue: 82,
-    saturation: 80,
-    lightness: 19,
-  },
-  background: {
-    hue: 82,
-    saturation: 46,
-    lightness: 66,
-  },
-  backgroundContrastText: {
-    hue: 82,
-    saturation: 70,
-    lightness: 22,
-  },
-};
+import FieldmarkImportDialog, {
+  FIELDMARK_COLOR,
+} from "./FieldmarkImportDialog";
 
 const StyledMenu = styled(Menu)(({ open }) => ({
   "& .MuiPaper-root": {
@@ -350,57 +315,12 @@ function CreateNew({ onClick }: CreateNewArgs): Node {
           }}
         />
       </StyledMenu>
-      <ThemeProvider theme={createAccentedTheme(FIELDMARK_COLOR)}>
-        <Dialog
-          open={fieldmarkOpen}
-          onClose={() => {
-            setFieldmarkOpen(false);
-            setAnchorEl(null);
-          }}
-        >
-          <AppBar position="relative" open={true}>
-            <Toolbar variant="dense">
-              <Typography variant="h6" noWrap component="h2">
-                Fieldmark
-              </Typography>
-              <Box flexGrow={1}></Box>
-              <Box ml={1}>
-                <AccessibilityTips
-                  supportsHighContrastMode
-                  elementType="dialog"
-                />
-              </Box>
-              <Box ml={1} sx={{ transform: "translateY(2px)" }}>
-                <HelpLinkIcon title="Fieldmark help" link="#" />
-              </Box>
-            </Toolbar>
-          </AppBar>
-          <Box sx={{ display: "flex", minHeight: 0 }}>
-            <DialogContent>
-              <Grid
-                container
-                direction="column"
-                spacing={2}
-                sx={{ height: "100%", flexWrap: "nowrap" }}
-              >
-                <Grid item>
-                  <Typography variant="h3">Import from Fieldmark</Typography>
-                </Grid>
-                <Grid item>
-                  <Typography variant="body2">
-                    Choose a Fieldmark notebook to import into Inventory. The
-                    new list container will be placed on your bench.
-                  </Typography>
-                  <Typography variant="body2">
-                    See <Link href="#">docs.fieldmark.au</Link> and our{" "}
-                    <Link href={"#"}>Fieldmark integration docs</Link> for more.
-                  </Typography>
-                </Grid>
-              </Grid>
-            </DialogContent>
-          </Box>
-        </Dialog>
-      </ThemeProvider>
+      <FieldmarkImportDialog
+        open={fieldmarkOpen}
+        onClose={() => {
+          setFieldmarkOpen(false);
+        }}
+      />
     </Box>
   );
 }
